#  OpenGL: Add Local Memory warmup shader for Nvidia