Here's a fun one.

Nvidia's OpenGL drivers see a significant performance uplift in shaders using local memory if the previous shader that was executed uses more local memory (sort of..)

To take advantage of this discovery, this change introduces a dummy shader that allocates the maximum allowed local memory, but returns immediately without executing. This tricks the driver into boosting the performance of the later executing shaders which use local memory.

Here's a local memory torture test homebrew used to test this: 
[textured_cube_compute.zip](https://github.com/yuzu-emu/yuzu/files/11861207/textured_cube_compute.zip)

In this homebrew, my performance went from 7 to 8 FPS, up to 40 to 42 FPS, matching Vulkan's performance.