@Kelebek1 2023-06-25
Does this not apply to Vulkan as well? All the local memory usage should be well beyond the graphics API. If this is OpenGL-specific, any idea why?