@ameerj 2023-06-25
This perf quirk does not affect Vulkan, only OpenGL, and it affects both GLSL and GLASM.
I can only guess at this point, but my assumption is that there's some aggressive optimization to limit lmem size in the GL driver, which gets reused in subsequent shader executions if they too use lmem.